+++ packages/components/organisms/f-checkout/CHANGELOG.md
@@ -6,9 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 Latest (add to next release)
 ------------------------------
-*February 22, 2021*
+*February 24, 2021*
 
 ### Changed
+- Restructured component object into page object model
+- Refactored component and accessibility tests
 - Extracted svgTransform to global test utils for reference by other packages